bpf_loader: fix same lint for the umpteenth time

+++ programs/bpf_loader/src/serialization.rs
@@ -404,7 +404,7 @@ fn serialize_parameters_aligned(
                 s.write::<u64>(borrowed_account.get_lamports().to_le());
                 s.write::<u64>((borrowed_account.get_data().len() as u64).to_le());
                 s.write_account(&mut borrowed_account)?;
-                s.write::<u64>((borrowed_account.get_rent_epoch() as u64).to_le());
+                s.write::<u64>((borrowed_account.get_rent_epoch()).to_le());
             }
             SerializeAccount::Duplicate(position) => {
                 s.write::<u8>(position as u8);
